requirements/requirements.txt
backcall==0.1.0
Click==7.0
decorator==4.4.1
Flask==1.1.1
grpcio==1.24.1
grpcio-tools==1.24.1
ipython==7.9.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
logconfig==0.4.0
logutils==0.3.5
MarkupSafe==1.1.1
meld3==2.0.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.6.0
Pygments==2.4.2
PyYAML==5.1.2
six==1.13.0
supervisor==4.0.4
traitlets==4.3.3
wcwidth==0.1.7
Werkzeug==0.16.0
